Update IDs from int to long in Models and Reports
Entity identifiers for Collection, Component, Metric, Alert, Label, and Agent have been updated from `int` to `long` across data models and reports to support larger identifier values. Related methods, parameters, and mappings have been adjusted accordingly. Additional code formatting, indentation, and cleanup improvements were also applied, with no further functional changes.

+++ ArcGISMonitorExcelReporterLib/Client/ArcGisMonitorQueryService.cs
@@ -562,7 +562,7 @@ namespace ArcGISMonitorExcelReporterLib.Client
         /// </code>
         /// </example>
         public async Task<QueryResponse<MetricFeature>> GetMetricTimeSeriesAsync(
-            IEnumerable<int> metricIds,
+            IEnumerable<long> metricIds,
             DateTimeOffset fromUtc,
             DateTimeOffset toUtc,
             string bucket = "observed_at:15m",
